requirements.txt
asgiref==3.6.0
Django==4.2
django-debug-toolbar==3.4.0
django-split-settings==1.2.0
djangorestframework==3.14.0
Faker==18.7.0
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
uWSGI==2.0.21
